requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
requests==2.31.0
python-dotenv==1.0.0
faiss-cpu==1.13.2
sentence-transformers==2.3.1
numpy>=2.0.0
pypdf2==3.0.1
python-docx==1.1.0